refactor(filesystem): inline embedding dimension default
