--- pageindex/filesystem/embedding_defaults.py
@@ -1,1 +1,0 @@
-DEFAULT_EMBEDDING_DIMENSIONS = 1024
