+++ pageindex/filesystem/hybrid_projection.py
@@ -10,7 +10,7 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Any
 
-from .embedding_defaults import DEFAULT_EMBEDDING_DIMENSIONS
+from .core import DEFAULT_EMBEDDING_DIMENSIONS
 from .semantic_index import SQLiteVecSemanticIndex, SemanticIndexError, SemanticSearchResult
